requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
seaborn
torch
torchvision
Pillow